requirements.txt
timm==0.4.12
transformers>=4.30
Pillow>=8.3.2
numpy>=1.21.1
opencv-python>=4.5.5.62
scipy>=1.8.0
scikit_image>=0.19.2
matplotlib>=3.4.3
scikit-learn
tensorboard
setuptools>=59.5.0

# Watermarking + training
torchvision

# FastAPI backend
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# CLIP-based authentication
open-clip-torch>=2.24.0

# JWT authentication
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# Property-based testing
hypothesis>=6.100.0

# YAML config
ruamel.yaml>=0.18.0